market/mysite/requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.6
asttokens==2.2.1
backcall==0.2.0
billiard==4.1.0
celery==5.3.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
cron-descriptor==1.4.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
executing==1.2.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
jupyter-client==8.3.0
jupyter-core==5.3.1
kombu==5.3.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
numpy==1.24.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.0
pylint==2.17.5
pymongo==3.12.3
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.1.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.2.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.16.2